requirements.txt
mysql-connector-python
requests
PyMuPDF
opencv-python-headless
numpy
python-dotenv
word2number
